Cosmetic changes: fixed button tooltips, moves config.json to data folder Issue #122 - Added a sample snapshot model to test files
