+++ src/components/gui/externalSearchBar.js
@@ -1,7 +1,7 @@
 import {Component, Input, NgModule} from '@angular/core';
 import {CommonModule} from "@angular/common";
 import {HttpClient, HttpClientModule} from '@angular/common/http';
-import {annotations} from "../config";
+import {annotations} from "../../data/config.json";
 
 @Component({
     selector: 'sciGraphSearch',
